more sampling

+++ src/components/Stats/Stats.hpp
@@ -5,13 +5,13 @@ namespace Dalea{
     struct Stats
     {
         uint64_t simple_splits;
-        uint64_t simple_split_time;
+        double simple_split_time;
         uint64_t traditional_splits;
-        uint64_t traditional_split_time;
+        double traditional_split_time;
         uint64_t complex_splits;
-        uint64_t complex_split_time;
+        double complex_split_time;
         uint64_t make_buddy;
-        uint64_t make_buddy_time;
+        double make_buddy_time;
 
         Stats() : simple_splits(0), traditional_splits(0), complex_splits(0), make_buddy(0),
                   simple_split_time(0),
